[Improvement] Init: move FallbackResponder to Http namespace
